Add files via upload
learnt the basics of support vector machines using Python and scikit-learn.  The dataset comes from the National Institute of Diabetes and Digestive and Kidney Diseases, and contains anonymized diagnostic measurements for a set of female patients.  Trained a support vector machine to predict whether a new patient has diabetes based on such measurements.